Add analyzder for multiline function declarations
This analyzer forces us to keep all fields on a single line, or to put each
on a separate line. By not having more than one field on a single line we
ensure that adding new arguments does not affect formatting of arguments before
or after the added argument, which keeps diffs small and simple.

Applies to both parameter and result lists.