Add analyzder for multiline function calls
This analyzer forces us to keep all arguments on a single line, or to put each
on a separate line. By not having more than one argument on a single line we
ensure that adding new arguments does not affect formatting of arguments before
or after the added argument, which keeps diffs small and simple.